Added more test code
The three A's
The three A's in the comments provide basic guidelines for your testing structure:

Arrange is where you declare any variables that the test might need. In this example, we must declare a calculator object to call its Add method.

Act is where you call the code that you want to test. In this step, you can insert parameters and exercise the code.

Assert is where you check if the outcome of the action is expected. You can add multiple assertions to any test. If one assertion fails, the test will fail.